Make package subpaths work properly
