+++ common/src/index.js
@@ -1,26 +1,9 @@
 const more = require("./more");
 const states = require("./states.json");
-
-/**
- * Parse a Newline-Delimited JSON (NDJSON) document.
- * @param {string} text
- * @returns {Array<any>}
- */
-function parseJsonLines(text) {
-  return text
-    .split("\n")
-    .filter(Boolean)
-    .map((line, index) => {
-      try {
-        return JSON.parse(line);
-      } catch (error) {
-        throw new SyntaxError(`Error parsing line ${index + 1}: ${line}`);
-      }
-    });
-}
+const utils = require("./utils");
 
 module.exports = {
-  parseJsonLines,
+  utils,
   more,
   states,
 };
